set up authorization for viewing consignment status- consolidated routes
