+++ server.js
@@ -16,8 +16,7 @@ app.engine(
 app.set("view engine", "handlebars");
 
 // Routes
-require("./routes/apiRoutes")(app);
-require("./routes/htmlRoutes")(app);
+require("./routes/routes")(app);
 
 var syncOptions = { force: false };
